separate GBK and GBKW
